bugfix: UnitPrice is type of decimal, which translates to float (instead of int) in python

+++ order_detail.py
@@ -4,4 +4,4 @@ from record_template import *
 
 class OrderDetail(Record):
     table_name = "OrderDetails"
-    fields = (("OrderID", int), ("ProductID", int), ("UnitPrice", int), ("Quantity", int), ("Discount", float))
+    fields = (("OrderID", int), ("ProductID", int), ("UnitPrice", float), ("Quantity", int), ("Discount", float))
